reorganizing cli directory
